requirements.txt
pipSQLAlchemy~=2.0.23
flask~=3.0.0